+++ package.json
@@ -23,12 +23,14 @@
   },
   "homepage": "https://github.com/csps/api#readme",
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.2",
+    "@types/bcryptjs": "^2.4.3",
     "@types/cors": "^2.8.13",
+    "@types/dotenv": "^8.2.0",
     "@types/express": "^4.17.17",
+    "@types/helmet": "^4.0.0",
     "@types/jsonwebtoken": "^9.0.2",
     "@types/mysql": "^2.15.21",
-    "@types/node": "^20.3.1",
+    "@types/node": "^20.3.2",
     "eslint": "^8.43.0",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.1.3"
